web/backend/pyproject.toml
[project]
name = "wp_embeddings_web_backend"
version = "0.1.0"
description = "Web application for 3D visualization of Wikipedia embeddings clusters"
readme = "../../README.md"
requires-python = ">=3.13"
packages=[".", "api", "app", "models", "services"]
dependencies = [
    "fastapi[standard]>=0.123.4",
    "pydantic>=2.12.3",
    "python-dotenv==1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "pytest==8.3.3",
    "pytest-asyncio>=1.3.0",
    "cachetools>=5.0.0",
]

[tool.flake8]
exclude = [".venv"]
max-line-length = 120

[dependency-groups]
dev = [
    "black>=25.11.0",
    "flake8-pyproject>=1.2.3",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.5",
    "vulture>=2.14",
]

[tool.ruff]
exclude = [
    ".venv"
]
line-length = 120

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
